Add command framework & preliminary version command
Changed inheritance from discord.Client to discord.ext.commands.Bot in
order to allow for command decorators. commands.Bot inherits from
discord.Client so the change is transparent

Added initial version checking framework based on the most recent git
tag in the home repository. Will fail if we're not on the master branch
because I don't have tags on develop.